+++ packages/connect-multichain/src/domain/index.ts
@@ -6,4 +6,3 @@ export * from './platform';
 export * from './store';
 export * from './ui';
 export * from './utils';
-export * from './versions';