+++ packages/connect-multichain/src/multichain/utils/analytics.ts
@@ -2,18 +2,14 @@
 /* eslint-disable @typescript-eslint/no-unused-vars -- Scope type used in JSDoc */
 import { getDappId } from '.';
 import type {
+  ConnectVersions,
   InvokeMethodOptions,
   MultichainOptions,
   PlatformType,
   Scope,
   StoreClient,
 } from '../../domain';
-import {
-  type ConnectVersions,
-  getPlatformType,
-  getVersions,
-  TransportType,
-} from '../../domain';
+import { getPlatformType, TransportType } from '../../domain';
 
 /**
  * Checks if an error represents a user rejection.
@@ -51,13 +47,12 @@ export async function getBaseAnalyticsProperties(
   options: MultichainOptions,
   storage: StoreClient,
 ): Promise<{
-  mmconnect_version: ConnectVersions;
+  mmconnect_version: ConnectVersions | Partial<ConnectVersions>;
   dapp_id: string;
   platform: PlatformType;
   integration_type: string;
   anon_id: string;
 }> {
-  const versions = getVersions();
   const dappId = getDappId(options.dapp);
   const platform = getPlatformType();
   const anonId = await storage.getAnonId();
@@ -66,7 +61,7 @@ export async function getBaseAnalyticsProperties(
       ?.integrationType ?? TransportType.UNKNOWN;
 
   return {
-    mmconnect_version: versions,
+    mmconnect_version: options.versions ?? {},
     dapp_id: dappId,
     platform,
     integration_type: integrationType,
@@ -87,14 +82,13 @@ export async function getWalletActionAnalyticsProperties(
   storage: StoreClient,
   invokeOptions: InvokeMethodOptions,
 ): Promise<{
-  mmconnect_version: ConnectVersions;
+  mmconnect_version: ConnectVersions | Partial<ConnectVersions>;
   dapp_id: string;
   method: string;
   integration_type: string;
   caip_chain_id: string;
   anon_id: string;
 }> {
-  const versions = getVersions();
   const dappId = getDappId(options.dapp);
   const anonId = await storage.getAnonId();
   const integrationType =
@@ -102,7 +96,7 @@ export async function getWalletActionAnalyticsProperties(
       ?.integrationType ?? 'unknown';
 
   return {
-    mmconnect_version: versions,
+    mmconnect_version: options.versions ?? {},
     dapp_id: dappId,
     method: invokeOptions.request.method,
     integration_type: integrationType,