+++ packages/connect-solana/vitest.config.ts
@@ -2,7 +2,13 @@
 import { resolve } from 'path';
 import { defineConfig } from 'vitest/config';
 
+import pkg from './package.json';
+
 export default defineConfig({
+  define: {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    __PACKAGE_VERSION__: JSON.stringify(pkg.version),
+  },
   test: {
     environment: 'jsdom',
     exclude: [